microservices/user-service/requirements.txt
asyncpg==0.20.1
databases[postgresql]==0.2.6
fastapi==0.52.0
SQLAlchemy==1.4
uvicorn==0.11.2
httpx==0.11.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# kafka-python==2.0.2
confluent-kafka==2.7.0